Rename module to data_import
